requirements.txt
# requirements.txt
pandas==1.5.3
numpy==1.24.3
yfinance==0.2.18
requests==2.31.0
beautifulsoup4==4.12.2
tweepy==4.14.0
praw==7.7.0
vaderSentiment==3.3.2
textblob==0.17.1
transformers==4.30.2
torch==2.0.1
scikit-learn==1.2.2
xgboost==1.7.5
tensorflow==2.12.0
streamlit==1.23.1
plotly==5.14.1
seaborn==0.12.2
matplotlib==3.7.1
ta==0.10.2
python-dotenv==1.0.0